+++ unnamed_engine/src/renderer/texture.rs
@@ -8,9 +8,11 @@ pub struct Texture {
   pub texture: wgpu::Texture,
   pub view: wgpu::TextureView,
   pub sampler: wgpu::Sampler,
+  pub bind_group: Option<wgpu::BindGroup>
 }
 
 impl Texture {
+  pub const DIFFUSE_FORMAT: wgpu::TextureFormat = wgpu::TextureFormat::Rgba8UnormSrgb;
   pub const DEPTH_FORMAT: wgpu::TextureFormat = wgpu::TextureFormat::Depth32Float;
   /// Load a texture from bytes
   pub fn from_bytes(
@@ -80,10 +82,57 @@ impl Texture {
       }
     );
 
-    Ok(Self { texture, view, sampler})
+    Ok(Self { texture, view, sampler, bind_group: None})
   }
 
-  pub fn create_depth_texture(device: &wgpu::Device, config: &wgpu::SurfaceConfiguration, label: &str) -> Self {
+  pub fn create_diffuse_texture(
+    device: &wgpu::Device,
+    config: &wgpu::SurfaceConfiguration,
+    label: &str,
+  ) -> Self {
+    let size = wgpu::Extent3d {
+      width: config.width,
+      height: config.height,
+      depth_or_array_layers: 1,
+    };
+    let desc = wgpu::TextureDescriptor {
+      label: Some(label),
+      size,
+      mip_level_count: 1,
+      sample_count: 1,
+      dimension: wgpu::TextureDimension::D2,
+      format: Self::DIFFUSE_FORMAT,
+      usage:
+        wgpu::TextureUsages::RENDER_ATTACHMENT |
+        wgpu::TextureUsages::TEXTURE_BINDING,
+      view_formats: &[],
+    };
+    let texture = device.create_texture(&desc);
+
+    let view = texture.create_view(&wgpu::TextureViewDescriptor::default());
+    let sampler = device.create_sampler(
+      &wgpu::SamplerDescriptor {
+        address_mode_u: wgpu::AddressMode::ClampToEdge,
+        address_mode_v: wgpu::AddressMode::ClampToEdge,
+        address_mode_w: wgpu::AddressMode::ClampToEdge,
+        mag_filter: wgpu::FilterMode::Linear,
+        min_filter: wgpu::FilterMode::Linear,
+        mipmap_filter: wgpu::FilterMode::Nearest,
+        compare: Some(wgpu::CompareFunction::LessEqual),
+        lod_min_clamp: 0.0,
+        lod_max_clamp: 100.0,
+        ..Default::default()
+      }
+    );
+
+    Self { texture, view, sampler, bind_group: None }
+  }
+
+  pub fn create_depth_texture(
+    device: &wgpu::Device,
+    config: &wgpu::SurfaceConfiguration,
+    label: &str,
+  ) -> Self {
     let size = wgpu::Extent3d {
       width: config.width,
       height: config.height,
@@ -119,6 +168,25 @@ impl Texture {
       }
     );
 
-    Self { texture, view, sampler }
+    Self { texture, view, sampler, bind_group: None }
+  }
+
+  pub fn set_bind_group(&mut self, device: &wgpu::Device, layout: &wgpu::BindGroupLayout) {
+    self.bind_group = Some(device.create_bind_group(
+      &wgpu::BindGroupDescriptor {
+        label: None,
+        layout,
+        entries: &[
+          wgpu::BindGroupEntry {
+            binding: 0,
+            resource: wgpu::BindingResource::TextureView(&self.view),
+          },
+          wgpu::BindGroupEntry {
+            binding: 1,
+            resource: wgpu::BindingResource::Sampler(&self.sampler),
+          },
+        ],
+      }
+    ));
   }
 }